cleanup and make thing clear
